\name{gentelman-package}
\alias{gentelman-package}
\alias{gentelman}
\docType{package}
\title{\packageTitle{gentelman}}
\description{\packageDescription{gentelman}}
\section{Package Content}{\packageIndices{gentelman}}
\author{\packageAuthor{gentelman}}
\section{Maintainer}{\packageMaintainer{gentelman}}
\keyword{package}
