% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{a placeholder}
\usage{
hello(txt = "world")
}
\arguments{
\item{txt}{character, name to say hello to}
}
\description{
longer description bla bla bla
}
\references{
Very important paper, P Sobczyk, 2020
}
