% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello_poznan}
\alias{hello_poznan}
\title{new function hello}
\usage{
hello_poznan(txt = "Pazur")
}
\arguments{
\item{txt}{character, name to say hello to}
}
\value{
character, hello string
}
\description{
new function hello
}
\examples{
hello_poznan("Pazur")
}
\seealso{
\code{\link[=hello]{hello()}}
}
